requirements.txt
pip==22.0.4
setuptools==62.1.0
keras==2.6.0
tensorflow==2.6.0
scikit-learn==0.24.2
numpy==1.21.2
pandas==1.3.2
PyYAML==5.4.1
